requirements.txt
absl-py==1.4.0
adjustText==0.8
aiocsv==1.3.1
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
astropy==6.0.0
astropy-iers-data==0.2024.3.4.0.30.17
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
autocorrect==2.6.1
azure-ai-ml==1.13.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.19.1
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.15.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bioinfokit==2.1.3
bleach==6.0.0
blinker==1.7.0
build==1.1.1
CacheControl==0.14.0
cachetools==5.3.0
casefy==0.1.7
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.0.1
Chroma==0.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
cleo==2.1.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
confluent-kafka==2.6.1
contourpy==1.0.6
crashtest==0.4.1
cryptography==42.0.5
cycler==0.11.0
dacite==1.8.1
dataclasses-avroschema==0.57.1
dataclasses-json==0.6.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
dulwich==0.21.7
durationpy==0.9
einsteinpy==0.4.0
emoji==2.14.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.2.0
faiss-cpu==1.10.0
fastapi==0.115.6
fastavro==1.9.4
fastjsonschema==2.19.1
filelock==3.12.4
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==23.1.21
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.3.3
fsspec==2023.9.2
gast==0.4.0
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.1
grpcio==1.69.0
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httpcore==1.0.3
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
ifaddr==0.2.0
imageio==2.36.1
imbalanced-learn==0.12.2
imblearn==0.0
implicit==0.7.2
importlib_metadata==7.0.2
importlib_resources==6.5.2
Inflector==3.1.1
iniconfig==2.0.0
installer==0.7.0
ipykernel==6.21.3
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.1
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.13.5
jupyter_core==5.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
keras==2.11.0
Keras-Preprocessing==1.1.2
keyring==24.3.1
kiwisolver==1.4.4
kubernetes==31.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langsmith==0.2.10
lazy_loader==0.4
libclang==15.0.6.1
llvmlite==0.42.0
lxml==5.0.1
Mako==1.3.2
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mdurl==0.1.2
mendel-development-tool==1.5.2
mistune==2.0.5
mlflow==2.11.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
msal==1.27.0
msal-extensions==1.1.0
msgpack==1.0.8
msrest==0.7.1
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook-as-pdf==0.5.0
notebook_shim==0.2.3
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.59.6
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencv-python==4.9.0.80
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt-einsum==3.3.0
orjson==3.10.14
overrides==7.7.0
packaging==23.2
pandas==1.5.2
pandoc==2.4
pandocfilters==1.5.0
paramiko==3.4.0
parso==0.8.3
patsy==0.5.4
peewee==3.17.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pkginfo==1.10.0
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.5.0
plumbum==1.8.3
ply==3.11
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.6.0
portalocker==2.8.2
posthog==3.7.5
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==5.29.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydash==7.0.5
pyee==11.1.1
pyerfa==2.0.1.1
Pygments==2.14.0
PyJWT==2.8.0
pymongo==4.6.2
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.0.9
PyPDF2==3.0.1
PyPika==0.48.9
pyppeteer==2.0.0
pyproject_hooks==1.0.0
pyreadline3==3.5.4
pyrsistent==0.19.3
pyTelegramBotAPI==4.16.1
pytest==8.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==1.1.6
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
querystring-parser==1.2.4
rapidfuzz==3.6.2
redlines==0.5.1
regex==2023.8.8
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-click==1.8.5
rsa==4.9
safetensors==0.3.3
scikit-fuzzy==0.4.2
scikit-image==0.25.0
scikit-learn==1.2.0
scipy==1.12.0
seaborn==0.13.0
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.41.3
statsmodels==0.14.0
strictyaml==1.7.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.1.0
tifffile==2024.12.12
tiktoken==0.8.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.4
torch==2.1.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.33.3
trove-classifiers==2024.3.3
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
umap==0.1.1
uri-template==1.2.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.25.1
waitress==3.0.0
watchfiles==1.0.4
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==3.0.1
widgetsnbextension==3.6.4
wordcloud==1.9.3
wrapt==1.14.1
yarl==1.9.2
yfinance==0.2.48
zeroconf==0.131.0
zipp==3.17.0
